refactor: rename utils.ts to index.ts
