+++ src/commands/callback.ts
@@ -7,7 +7,7 @@
 import {FolderSettings, RepoFrontmatter, Repository} from "../settings/interface";
 import GithubPublisher from "../main";
 import {checkRepositoryValidity, isShared} from "../utils/data_validation_test";
-import {createLink, getRepoFrontmatter} from "../utils/utils";
+import {createLink, getRepoFrontmatter} from "../utils";
 import i18next from "i18next";
 import {Command, Notice } from "obsidian";
 import {purgeNotesRemote, shareOneNote} from "./commands";
